move poll progress participation indication formatting to VoteCount DTO
trying to decouple text formatting with keeping track of total votes in regard to group size caused code duplication and confusion
