+++ src/main/kotlin/com/wire/apps/polls/services/StatsFormattingService.kt
@@ -1,13 +1,11 @@
 package com.wire.apps.polls.services
 
 import com.wire.apps.polls.dao.PollRepository
-import com.wire.apps.polls.dto.VoteCount
 import com.wire.apps.polls.dto.common.Text
 import mu.KLogging
 import pw.forst.katlib.newLine
 import pw.forst.katlib.whenNull
 import kotlin.math.min
-import kotlin.math.roundToInt
 
 class StatsFormattingService(
     private val repository: PollRepository
@@ -105,7 +103,7 @@ class StatsFormattingService(
     private fun prepareTitle(body: String) = "$TITLE_PREFIX${body}\"*"
 }
 
-private class VoteDisplay(val voted: Int, val outOf: Int) {
+internal class VoteDisplay(val voted: Int, val outOf: Int) {
     private companion object {
         const val NOT_VOTE = "⚪"
         const val VOTE = "🟢"
@@ -131,22 +129,5 @@ private data class VotingOption(
         val voteDisplay = VoteDisplay(votingUsers, max)
 
         return "$voteDisplay $style$option$style ($votingUsers)"
-    }
-}
-
-object VotingCount {
-    const val BLOCKS = 10
-    const val PERCENTAGE_FACTOR = 100
-
-    fun new(): String {
-        return VoteCount.initial().update()
-    }
-
-    fun VoteCount.update(): String {
-        val percent = if (totalMembers > 0) votesCast.toDouble() / totalMembers else 0.0
-        val voteDisplay = VoteDisplay((percent * BLOCKS).roundToInt(), BLOCKS)
-        val percentDisplay = (percent * PERCENTAGE_FACTOR).roundToInt()
-
-        return "$voteDisplay $percentDisplay%"
     }
 }
